% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_and_slice}
\alias{wrangle_and_slice}
\title{Wrangle input data for format required by simulation}
\usage{
wrangle_and_slice(start_year, end_year, inndata, focal_species,
  geoselect_native)
}
\arguments{
\item{start_year}{Year of staring recording as introductions}

\item{end_year}{Year of ending recording as introductions}

\item{inndata}{data-frame with inndata comming from "get_inndata"}

\item{focal_species}{The latin (canonical) name of the species}

\item{geoselect_native}{A spatial.polygon.dataframe indicating the native distribution range of the species in epsg:4326}
}
\value{
A data.frame
}
\description{
Wrangle input data for format required by simulation
}
\examples{

inndata <- readRDS("./testdata/view_occurrence_by_event.rds")
geoselect_native <- readRDS("./testdata/geoselect_native_Esox_lucius.rds")
focal_species <- "Esox lucius"
mynewdata_aslist <- wrangle_and_slice(inndata=inndata,start_year=1970,end_year=2017)
mynewdata <- outdata[["data"]]

}
